# Add a missing index + small DB optimizations