Initial migration time: ~100 seconds -> 95 seconds
Re-migration time: ~95 seconds -> 75 seconds

[noissue]