@dralley 2020-03-18
New graphs:

Initial migration:

![initial_migration](https://user-images.githubusercontent.com/2751996/77008369-6adc5400-693c-11ea-9e81-8faa81aa5835.png)

Re-migration:

![re_migration](https://user-images.githubusercontent.com/2751996/77008386-729bf880-693c-11ea-84d6-6784d959c73d.png)