@dralley 2020-03-19
Updated cprofile graphs (including https://github.com/pulp/pulpcore/pull/597 as well):

Initial migration:

![initial_migration](https://user-images.githubusercontent.com/2751996/77023721-04b6f780-6963-11ea-9e76-32716c676ed4.png)

Re migration:

![re_migration](https://user-images.githubusercontent.com/2751996/77023715-008ada00-6963-11ea-8f70-91ec3b0545a0.png)