@goosemania 2020-03-19
@dralley , do we need a migration to build indexes?