@dralley 2020-03-19
@goosemania Yes, indexes require a migration